refactor: flatten models folder structure
