+++ simba/analog_discovery/fc_layers_analog_discovery.py
@@ -2,7 +2,7 @@ import numpy as np
 import torch
 from tqdm import tqdm
 
-from simba.core.models.ordinal.embedder_multitask import EmbedderMultitask
+from simba.core.models.embedder_multitask import EmbedderMultitask
 
 
 class FcLayerAnalogDiscovery:
