+++ simba/core/models/__init__.py
@@ -1,7 +1,7 @@
 """Neural network models for SIMBA."""
 
-from simba.core.models.transformers.embedder import Embedder
-from simba.core.models.transformers.spectrum_transformer_encoder_custom import (
+from simba.core.models.embedder import Embedder
+from simba.core.models.spectrum_transformer_encoder_custom import (
     SpectrumTransformerEncoderCustom,
 )
